Wew that's a lot of changes. Because I'm too lazy to commit super incomplete work. But this is still incomplete...
Alright. I've been working on this for the better part of this summer
now. And a bunch of work is coming with this commit...
1. Overhauled Input System: There way we read client inputs is slightly
different now. Basically much prettier and also a bit more intuitive, if
not too verbose. It also plays along nicely with the new Action system.
Just need a way to save the inputs, which is coming soon...

2. New Action System: Insstead of handwaving and hoping things turn out
alright on the server side, the Action system introduces some
much-needed concreteness to your... actions. Alongside the Input system,
the Action system allows for new possibilities, but since I'm new to all
this, I still give the game programmer plenty of control over Actions. I
haven't covered the whole thing in a black box. Yet.

3. Minor Infrastructure Changes: Yep. Nowadays who can't afford the
capacity required for 32 bit ints? Honestly I'm not sure why I tried to
hard to save a byte or two here and there. Most integer numbers should
now be ints.

4. Reverted 2-Tuple System: Also not too sure what I was thinking. Like
Vector2 by itself is honestly sufficient for all purposes... no real
point breaking it down into Vector2 Point2 and Mutable Tuple2s...

5. Graph Utilities: Located under the math utilities comes graph
utilities. At the moment there is only pathfinding, and even that is
incomplete, but it's a start. Also on the docket is revamping the
LevelTileds to work with the new graph utilities, because aren't tiled
levels basically graphs?

Whew. Lots of work all rolled into one. I hope I didn't delete code I
needed, but then again it's Git. Gonna be around. Somewhere...

+++ src/engine/Game.java
@@ -71,7 +71,7 @@ public abstract class Game {
 	/**
 	 * The next available Player Number
 	 */
-	public short nextPlayerNumber = 0;
+	public int nextPlayerNumber = 0;
 	
 	/**
 	 * Creates a new {@code Game}
@@ -140,7 +140,7 @@ public abstract class Game {
 	public Player getNewPlayerInstance() {
 		Class<? extends Player> cls = this.getPlayerClass();
 		try {
-			Constructor<? extends Player> cst = cls.getConstructor(Game.class, short.class);
+			Constructor<? extends Player> cst = cls.getConstructor(Game.class, int.class);
 			Player p = cst.newInstance(this, this.nextPlayerNumber++);
 			return p;
 		} catch (Exception e) {
@@ -159,7 +159,7 @@ public abstract class Game {
 	public Player getNewPlayerInstance(String name) {
 		Class<? extends Player> cls = this.getPlayerClass();
 		try {
-			Constructor<? extends Player> cst = cls.getConstructor(Game.class, short.class, String.class);
+			Constructor<? extends Player> cst = cls.getConstructor(Game.class, int.class, String.class);
 			Player p = cst.newInstance(this, this.nextPlayerNumber++, name);
 			return p;
 		} catch (Exception e) {
