+++ src/engine/event/game/ChatEvent.java
@@ -4,11 +4,11 @@ import engine.event.Event;
 
 public class ChatEvent extends Event {
 	
-	public short pnum;
+	public int pnum;
 	
 	public String msg;
 	
-	public ChatEvent(short pnum, String msg) {
+	public ChatEvent(int pnum, String msg) {
 		this.pnum = pnum;
 		this.msg = msg;
 	}
