+++ src/engine/example/PacketPlayerScoreNIO.java
@@ -7,30 +7,31 @@ import engine.client.Client;
 import engine.networknio.packet.PacketNIO;
 import engine.server.Server;
 
-public class PacketPlayerScoreNIO extends PacketNIO{
+public class PacketPlayerScoreNIO extends PacketNIO {
 	
-	public short pnum;
+	
+	public int pnum;
 	
 	public int score;
 	
 	public PacketPlayerScoreNIO() {
-	
+		
 	}
 	
-	public PacketPlayerScoreNIO(short pnum, int score) {
+	public PacketPlayerScoreNIO(int pnum, int score) {
 		this.pnum = pnum;
 		this.score = score;
 	}
 	
 	@Override
 	public void writePacketData(ByteBuffer buff) throws IOException {
-		buff.putShort(this.pnum);
+		buff.putInt(this.pnum);
 		buff.putInt(this.score);
 	}
 	
 	@Override
 	public void readPacketData(ByteBuffer buff) throws IOException {
-		this.pnum = buff.getShort();
+		this.pnum = buff.getInt();
 		this.score = buff.getInt();
 	}
 	
@@ -40,7 +41,7 @@ public class PacketPlayerScoreNIO extends PacketNIO{
 	}
 	
 	@Override
-	public void processServer(short player, Server s) {
+	public void processServer(int player, Server s) {
 	}
 	
 }
