+++ src/engine/networknio/packet/PacketPlayer.java
@@ -43,7 +43,7 @@ public class PacketPlayer extends PacketNIO {
 	}
 	
 	@Override
-	public void processServer(short player, Server s) {
+	public void processServer(int player, Server s) {
 	}
 	
 }
